+++ tools/generate_authors.py
@@ -56,7 +56,8 @@ def main(repos=None, output_path=None):
                  'https://github.com/Qiskit/qiskit-aer',
                  'https://github.com/Qiskit/qiskit-aqua',
                  'https://github.com/Qiskit/qiskit-ignis',
-                 'https://github.com/Qiskit/qiskit-chemistry']
+                 'https://github.com/Qiskit/qiskit-chemistry',
+                 'https://github.com/Qiskit/qiskit-jku-provider']
     if not output_path:
         output_path = 'AUTHORS'
     authors = []