+++ CHANGELOG.rst
@@ -15,8 +15,55 @@ The format is based on `Keep a Changelog`_.
   - **Fixed**: for any bug fixes.
   - **Security**: in case of vulnerabilities.
 
+
 `UNRELEASED`_
 =============
+
+
+`0.10.1`_ - 2019-05-07
+======================
+
+Changed
+-------
+
+- Updated the ``qiskit-ibmq-provider`` dependency to 0.2.2.
+
+
+`0.10.0`_ - 2019-05-06
+======================
+
+Changed
+-------
+
+- Updated the ``qiskit-ibmq-provider`` dependency to 0.2.1 (#244).
+
+
+`0.9.0`_ - 2019-05-02
+=====================
+
+
+Added
+-----
+
+- Added the qiskit-ibmq-provider package to the metapackage (#227).
+- Added qiskit-aqua and qiskit-chemistry to the set of installed packages (#185)
+
+Changed
+-------
+
+- The qiskit-terra version increased to the next feature release 0.8.0 (#227).
+- The qiskit-aer version increased to the next feature release 0.2.0 (#227).
+- The qiskit-ignis version increased to the next release 0.1.1 (#227).
+
+
+`0.8.1`_ - 2019-05-01
+=====================
+
+Changed
+-------
+
+- Increased the qiskit-terra version to v0.7.2 which includes a fix for the
+  schema validation of backend configuration in Qiskit/qiskit-terra#2258 (#217)
 
 
 `0.8.0`_ - 2019-03-05
@@ -65,7 +112,11 @@ Added
 
 - First release, includes qiskit-terra and qiskit-aer
 
-.. _UNRELEASED: https://github.com/Qiskit/qiskit-terra/compare/0.8.0...HEAD
+.. _UNRELEASED: https://github.com/Qiskit/qiskit-terra/compare/0.10.1...HEAD
+.. _0.10.1: https://github.com/Qiskit/qiskit/compare/0.10.0...0.10.1
+.. _0.10.0: https://github.com/Qiskit/qiskit/compare/0.9.0...0.10.0
+.. _0.9.0: https://github.com/Qiskit/qiskit/compare/0.8.1...0.9.0
+.. _0.8.1: https://github.com/Qiskit/qiskit/compare/0.8.0...0.8.1
 .. _0.8.0: https://github.com/Qiskit/qiskit/compare/0.7.3...0.8.0
 .. _0.7.3: https://github.com/Qiskit/qiskit/compare/0.7.2...0.7.3
 .. _0.7.2: https://github.com/Qiskit/qiskit/compare/0.7.1...0.7.2